Fitting a Model in lavaan or in survey

The function BIFIE.lavaan.survey fits a structural equation model in lavaan using the lavaan.survey package. Currently, only maximum likelihood estimation for normally distributed data is available.

The function BIFIE.survey fits a model defined in the survey package.

BIFIE.lavaan.survey(lavmodel, svyrepdes, lavaan_fun="sem",
    lavaan_survey_default=FALSE, fit.measures=NULL, ...)

# S3 method for BIFIE.lavaan.survey
summary(object, ...)

# S3 method for BIFIE.lavaan.survey
coef(object,...)

# S3 method for BIFIE.lavaan.survey
vcov(object,...)

BIFIE.survey(svyrepdes, survey.function, ...)

# S3 method for BIFIE.survey
summary(object, digits=3, ...)

# S3 method for BIFIE.survey
coef(object,...)

# S3 method for BIFIE.survey
vcov(object,...)

Arguments

lavmodel
Model string in lavaan syntax

svyrepdes
Replication design object of class BIFIEdata or replication design object from survey package (generated by BIFIEdata2svrepdesign or survey::svrepdesign)

lavaan_fun
Estimation funcion in lavaan. Can be "lavaan", "sem", "cfa" or "growth".

lavaan_survey_default
Logical indicating whether the lavaan.survey package should be used for statistical inference for multiply imputed datasets.

object
Object of class BIFIE.by

fit.measures
Optional vector of fit measures used in lavaan::fitMeasures function

...
Further arguments to be passed

survey.function
Function from the survey package

digits
Number of digits after decimal

Value

For BIFIE.lavaan.survey a list with following entries

lavfit

Object of class lavaan

fitstat

Fit statistics from lavaan

See also

Examples